split RewritingNamer into its own file, commentary
