+++ namer/core/src/main/scala/io/buoyant/hostport.scala
@@ -1,30 +1,14 @@
 package io.buoyant
 
-import com.twitter.finagle.{Name, NameTree, Namer, Path}
-import com.twitter.util.{Activity, Try}
+import com.twitter.finagle.Path
 
 /**
- * A helper
+ * Extracts a host and port from a "host:port" string.
+ *
+ * Does not not support IPv6 host IPs (because ipv6 notation doesn't
+ * work in Paths as-is, due to bracket characters).
  */
-trait RewritingNamer extends Namer {
-  protected[this] def rewrite(orig: Path): Option[Path]
-
-  def lookup(path: Path): Activity[NameTree[Name]] =
-    Activity.value(path match {
-      case path if path.size > 0 =>
-        rewrite(path) match {
-          case Some(path) => NameTree.Leaf(Name.Path(path))
-          case None => NameTree.Neg
-        }
-      case _ => NameTree.Neg
-    })
-}
-
 private object HostColonPort {
-  // do very coarse host matching so that we can support dns names,
-  // ipv4, and ipv6 without going crazy.
-  val hostRE = """^[A-Za-z0-9.:_-]+$""".r
-
   object PortStr {
     def unapply(s: String): Option[Int] = {
       val port = try s.toInt catch { case _: java.lang.NumberFormatException => -1 }
@@ -58,7 +42,7 @@ class hostportPfx extends RewritingNamer {
 /**
  * A rewriting namer that accepts names in the form:
  *
- *   /host:port/pfx/etc
+ *   /pfx/host:port/etc
  *
  * and rewrites to:
  *
